Remove components/messages and move page to server component
