+++ README.md
@@ -279,7 +279,6 @@ fixes and corrections above.
 - [components/FeaturedProperties.jsx](components/FeaturedProperties.jsx)
 - [app/properties/[id]/page.jsx](app/properties/[id]/page.jsx)
 - [app/properties/search-results/page.jsx](app/properties/search-results/page.jsx)
-- [components/Messages.jsx](components/Messages.jsx)
 
 The above changes also depend on marking
 [PropertyImages](components/PropertyImages.jsx) and
@@ -296,6 +295,12 @@ in our [utils/requests.js](utils/requests.js) module.
 We can remove our **app/api/properties/search/route.js** as it's no longer used.
 
 We can remove the **GET** function from our [app/api/messages/route.js](app/api/messages/route.js) as it's no longer used.
+
+### Remove components/Messages.jsx and change the page to a server component
+
+The [app/messages/page.jsx](app/message/page.jsx) component just returns the
+`Messages` component so it seems simpler to just move the contents of `Messages`
+to the page component and remove `Messages`. Additionally [app/messages/page.jsx](app/message/page.jsx) can be made a server component.
 
 ## License
